ecommerce_tool/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
celery==5.3.4
cffi==1.17.1
cryptography==44.0.2
Django==4.2.19
django-celery-beat==2.5.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
django-redis>=5.4.0
et-xmlfile==2.0.0
gunicorn==21.2.0
jdcal==1.4.1
lxml==5.3.0
mongoengine==0.29.1
numpy>=1.26.0
openpyxl==3.1.5
pandas==2.0.3
pdfplumber==0.11.5
pycparser==2.22
PyJWT==2.8.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
setuptools>=67.0.0
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing-extensions==4.12.2
tzdata==2025.1
emoji==2.12.1